+++ thrust/system/cuda/detail/scan.h
@@ -669,7 +669,7 @@ namespace __scan {
                                            num_tiles);
 
     size_t allocation_sizes[2] = {0, vshmem_size};
-    status = ScanTileState::AllocationSize(num_tiles, allocation_sizes[0]);
+    status = ScanTileState::AllocationSize(static_cast<int>(num_tiles), allocation_sizes[0]);
     CUDA_CUB_RET_IF_FAIL(status);
 
     void* allocations[2] = {NULL, NULL};
@@ -686,7 +686,7 @@ namespace __scan {
     }
     
     ScanTileState tile_state;
-    status = tile_state.Init(num_tiles, allocations[0], allocation_sizes[0]);
+    status = tile_state.Init(static_cast<int>(num_tiles), allocations[0], allocation_sizes[0]);
     CUDA_CUB_RET_IF_FAIL(status);
 
     char *vshmem_ptr = vshmem_size > 0 ? (char*)allocations[1] : NULL;
